+++ src/helpers/dom-exists.js
@@ -1,0 +1,10 @@
+/**
+ * Check if the dom exists - useful for some server-side-rendering applications
+ */
+const domExists = () => !!(
+  typeof window !== 'undefined'
+    && window.document
+    && window.document.createElement
+);
+
+export default domExists;