+++ src/index.js
@@ -6,3 +6,4 @@ export { default as Optimize } from './optimize';
 export { default as ExperimentOpacifier } from './experiment-opacifier';
 export { default as Globals } from './experiment-globals';
 export { default as withExperiment } from './with-experiment';
+export { default as startMixpanelExperiment } from './start-mixpanel-experiment';