+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][]
+### Added
+- mixpanel experiment start function
+
+### Updated
+- yarn upgrade
 
 ## [1.0.9][] - 2019-10-25
 - Replace deprecated `componentWillMount()` with `componentDidMount()`